python_services/packages/text_extract_ideas/requirements.txt
langchain>=0.1.0
langchain-google-vertexai>=1.0.0
langgraph>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# PDF Processing
PyMuPDF>=1.23.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Progress & Logging
tqdm>=4.66.0
pyyaml>=6.0.0

# Async/Parallel Processing
aiohttp>=3.9.0
